requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
djangorestframework==3.7.7
idna==2.6
psycopg2==2.7.4
PyJWT==1.6.1
PySocks==1.6.8
pytz==2018.3
requests==2.18.4
six==1.11.0
twilio==6.10.5
urllib3==1.22
